mechanically move evaluate from base.ts into pyrepl.ts
