+++ pyscriptjs/src/components/pyrepl.ts
@@ -5,12 +5,16 @@ import { Compartment, StateCommand } from '@codemirror/state';
 import { keymap } from '@codemirror/view';
 import { defaultKeymap } from '@codemirror/commands';
 import { oneDarkTheme } from '@codemirror/theme-one-dark';
-import { getAttribute, addClasses, htmlDecode } from '../utils';
+import { getAttribute, addClasses, removeClasses,
+         ensureUniqueId, htmlDecode } from '../utils';
 import { BaseEvalElement } from './base';
 import type { Runtime } from '../runtime';
+import { pyExecDontHandleErrors } from '../pyexec';
 import { getLogger } from '../logger';
 
 const logger = getLogger('py-repl');
+
+let Element;
 
 export function make_PyRepl(runtime: Runtime) {
 
@@ -148,6 +152,45 @@ export function make_PyRepl(runtime: Runtime) {
             this.editor.focus();
             logger.debug(`element ${this.id} successfully connected`);
         }
+
+        async evaluate(runtime: Runtime): Promise<void> {
+            this.preEvaluate();
+
+            let source: string;
+            try {
+                source = this.source ? await this.getSourceFromFile(this.source)
+                                     : this.getSourceFromElement();
+
+                // XXX we should use pyExec and let it display the errors
+                await pyExecDontHandleErrors(runtime, source, this);
+
+                removeClasses(this.errorElement, ['py-error']);
+                this.postEvaluate();
+            } catch (err) {
+                logger.error(err);
+                try{
+                    if (Element === undefined) {
+                        Element = <Element>runtime.globals.get('Element');
+                    }
+                    const out = Element(this.errorElement.id);
+
+                    addClasses(this.errorElement, ['py-error']);
+                    out.write.callKwargs(err.toString(), { append: this.appendOutput });
+                    if (this.errorElement.children.length === 0){
+                        this.errorElement.setAttribute('error', '');
+                    }else{
+                        this.errorElement.children[this.errorElement.children.length - 1].setAttribute('error', '');
+                    }
+
+                    this.errorElement.hidden = false;
+                    this.errorElement.style.display = 'block';
+                    this.errorElement.style.visibility = 'visible';
+                } catch (internalErr){
+                    logger.error("Unnable to write error to error element in page.")
+                }
+
+            }
+        } // end evaluate
 
         preEvaluate(): void {
             this.setOutputMode("replace");
